+++ packages/desktop/src/components/auto-complete/auto-complete.vue
@@ -74,11 +74,11 @@
 import { useFormFallbackProps } from '@veltra/compositions'
 import { vRipple } from '@veltra/directives'
 import { bem, FORM_EMPTY_CONTENT, scrollIntoContainerView } from '@veltra/utils'
+import { injectFormContext } from '@veltra/utils'
 import { computed, shallowRef, watch } from 'vue'
 
 import type { AutoCompleteEmits, AutoCompleteProps, _AutoCompleteExposed } from '../../types'
 import type { DropdownExposed, ScrollExposed } from '../../types'
-import { injectFormContext } from '../../utils/form-context'
 import { UDropdown } from '../dropdown'
 import { UInput } from '../input'
 import { UScroll } from '../scroll'