+++ packages/desktop/src/components/checkbox/checkbox-button.vue
@@ -15,10 +15,10 @@
 <script lang="ts" setup>
 import { useFormFallbackProps } from '@veltra/compositions'
 import { bem, FORM_EMPTY_CONTENT } from '@veltra/utils'
+import { injectFormContext } from '@veltra/utils'
 import { computed } from 'vue'
 
 import type { CheckboxButtonProps, CheckboxButtonEmits } from '../../types'
-import { injectFormContext } from '../../utils/form-context'
 import { UTag } from '../tag'
 
 defineOptions({