+++ packages/desktop/src/components/form-item/form-item.vue
@@ -35,10 +35,10 @@
 <script lang="tsx" setup>
 import { useConfig, useFallbackProps } from '@veltra/compositions'
 import { bem, withUnit } from '@veltra/utils'
+import { injectFormContext } from '@veltra/utils'
 import { type CSSProperties, computed } from 'vue'
 
 import type { FormItemProps, ComponentSize } from '../../types'
-import { injectFormContext } from '../../utils/form-context'
 import { UGridItem } from '../grid'
 import { UTip } from '../tip'
 import { formItemCls as cls } from './helper'