+++ packages/desktop/src/components/password-input/password-input.vue
@@ -30,10 +30,10 @@ import { o } from '@cat-kit/core'
 import { useFormFallbackProps } from '@veltra/compositions'
 import { Hide, View } from '@veltra/icons/normal'
 import { bem } from '@veltra/utils'
+import { injectFormContext } from '@veltra/utils'
 import { computed, nextTick, shallowRef } from 'vue'
 
 import type { PasswordInputProps } from '../../types'
-import { injectFormContext } from '../../utils/form-context'
 import { UIcon } from '../icon'
 import { UInput } from '../input'