+++ packages/desktop/src/components/pop-confirm/pop-confirm.vue
@@ -35,10 +35,10 @@
 import { useFormFallbackProps } from '@veltra/compositions'
 import { QuestionFilled } from '@veltra/icons/normal'
 import { bem } from '@veltra/utils'
+import { injectFormContext } from '@veltra/utils'
 import { ref } from 'vue'
 
 import type { PopConfirmProps, PopConfirmEmits } from '../../types'
-import { injectFormContext } from '../../utils/form-context'
 import { UButton } from '../button'
 import { UIcon } from '../icon'
 import { UTip } from '../tip'