+++ packages/desktop/src/components/slider/slider.vue
@@ -22,10 +22,10 @@
 <script lang="ts" setup generic="T extends number | [number, number]">
 import { useFormFallbackProps, useUserAction } from '@veltra/compositions'
 import { bem } from '@veltra/utils'
+import { injectFormContext } from '@veltra/utils'
 import { computed, provide, watch } from 'vue'
 
 import type { SliderProps, SliderEmits } from '../../types'
-import { injectFormContext } from '../../utils/form-context'
 import { sliderContextKey } from './di'
 import SliderThumb from './slider-thumb.vue'
 import { useSlider } from './use-slider'