+++ packages/desktop/src/components/tree/tree.vue
@@ -33,11 +33,11 @@
 <script lang="ts" setup>
 import { useFormFallbackProps, useVirtualizer } from '@veltra/compositions'
 import { bem, nextFrame, scrollIntoContainerView } from '@veltra/utils'
+import { injectFormContext } from '@veltra/utils'
 import { computed, provide, shallowRef, useSlots, watch, watchEffect, type VNode } from 'vue'
 
 import type { TreeProps, TreeEmit, _TreeExposed } from '../../types'
 import type { ScrollExposed } from '../../types'
-import { injectFormContext } from '../../utils/form-context'
 import { UEmpty } from '../empty'
 import { UScroll } from '../scroll'
 import { TreeDIKey, type TreeConText, type TreeSlotsScope } from './di'