+++ packages/compositions/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @veltra/compositions
 
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @veltra/utils@1.1.3
+
 ## 1.1.2
 
 ### Patch Changes